LPS-76024 Log handled exception
Source formatter requires a caught exception to be processed directly in
some way. Ideally, we would use a fetch method to retrieve a folder,
which would return null instead of throwing an exception. However, no
such method exists in the PortletFileRepositoryUtil interface, so we are
constrained to relying on exception handling to determine when the
requested folder doesn't exist.